+++ it/uk/gov/hmrc/tai/integration/PutWorkingFromHomeEmployeeExpenses.scala
@@ -32,7 +32,8 @@ class PutWorkingFromHomeEmployeeExpenses extends IntegrationSpec {
 
   def request = FakeRequest(POST, apiUrl)
     .withJsonBody(postRequest)
-    .withHeaders(HeaderNames.authorisation -> bearerToken)
+    .withHeaders(HeaderNames.authorisation -> bearerToken,
+      HeaderNames.xSessionId -> "sessionId")
 
   val iabdType = 59
   val desIabdsUrl = s"/pay-as-you-earn/individuals/$nino/iabds/$year/$iabdType"
